requirements.txt
alembic==1.5.2
aniso8601==8.1.0
attrs==20.3.0
automium==0.2.6
automium-web==0.1.1
click==7.1.2
coverage==5.4
Flask==1.1.0
Flask-Admin==1.5.3
Flask-Migrate==2.5.2
flask-restplus==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
legacy==0.1.6
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.0.0rc8
mock==4.0.3
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
pytest-flask==1.1.0
pytest-it==0.1.4
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==0.16.1
WTForms==2.3.3
